Dodana Post metoda ki vrne list prs_spot

+++ resPrsSpotApi/Classes/prs_Spot.cs
@@ -2,7 +2,7 @@
 {
     public class prs_Spot
     {
-        public double maticnaEnotaField { get; set; }
+        public int maticnaEnotaField { get; set; }
         public string popolnoImeField { get; set; }
         public string kratkoImeField { get; set; }
     }
